user_(action) method for Comment|Product|User Controller + api update

+++ app/Http/Controllers/ProductController.php
@@ -1,10 +1,12 @@
 <?php
 namespace App\Http\Controllers;
 
+use App\Http\Auth;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
+use App\Models\Category;
 use Illuminate\Support\Str;
 
 
@@ -20,7 +22,49 @@ class ProductController extends Controller
     	$products = Product::where('id', '>', -1)
         ->orderBy('created_at', 'desc');
 
-        if ($request->input('page') == null || 
+        if ($request->input('page') == null ||
+            $request->input('page') == '') {
+            $products = $products->get();
+        } else {
+            $products = $products->paginate();
+        }
+
+        $data = [
+            'success' => true,
+            'products' => $products
+        ];
+
+        return response()->json($data);
+    }
+
+    public function category_index(Request $request, Category $category)
+    {
+    	$products = Product::where('category_id', $category->id)
+        ->orderBy('created_at', 'desc');
+
+        if ($request->input('page') == null ||
+            $request->input('page') == '') {
+            $products = $products->get();
+        } else {
+            $products = $products->paginate();
+        }
+
+        $data = [
+            'success' => true,
+            'products' => $products
+        ];
+
+        return response()->json($data);
+    }
+
+    public function user_index(Request $request)
+    {
+        $user = Auth::getUser($request, Auth::USER);
+
+    	$products = Product::where('user_id', $user->id)
+        ->orderBy('created_at', 'desc');
+
+        if ($request->input('page') == null ||
             $request->input('page') == '') {
             $products = $products->get();
         } else {
@@ -58,7 +102,7 @@ class ProductController extends Controller
         $product = new Product;
 
         $product->nom = $validated['nom'] ?? null;
-		$product->slug = $validated['slug'] ?? null;
+		$product->slug = Str::slug($validated['slug']);
 		$product->description = $validated['description'] ?? null;
 		$product->prix = $validated['prix'] ?? null;
 		$product->type_paiement = $validated['type_paiement'] ?? null;
@@ -68,14 +112,43 @@ class ProductController extends Controller
 		$product->category_id = $validated['category_id'] ?? null;
 		$product->municipality_id = $validated['municipality_id'] ?? null;
 		$product->user_id = $validated['user_id'] ?? null;
-		
+
         $product->save();
 
         $data = [
             'success'       => true,
             'product'   => $product
         ];
-        
+
+        return response()->json($data);
+    }
+
+    public function user_store(StoreProductRequest $request)
+    {
+        $user = Auth::getUser($request, Auth::USER);
+        $validated = $request->validated();
+
+        $product = new Product;
+
+        $product->nom = $validated['nom'] ?? null;
+		$product->slug = Str::slug($validated['slug']);
+		$product->description = $validated['description'] ?? null;
+		$product->prix = $validated['prix'] ?? null;
+		$product->type_paiement = $validated['type_paiement'] ?? null;
+		$product->type = $validated['type'] ?? null;
+		$product->display_img_url_list = $validated['display_img_url_list'] ?? null;
+		$product->images_url_list = $validated['images_url_list'] ?? null;
+		$product->category_id = $validated['category_id'] ?? null;
+		$product->municipality_id = $validated['municipality_id'] ?? null;
+		$product->user_id = $user->id;
+
+        $product->save();
+
+        $data = [
+            'success'       => true,
+            'product'   => $product
+        ];
+
         return response()->json($data);
     }
 
@@ -118,7 +191,7 @@ class ProductController extends Controller
         $validated = $request->validated();
 
         $product->nom = $validated['nom'] ?? null;
-		$product->slug = $validated['slug'] ?? null;
+		$product->slug = Str::slug($validated['slug']);
 		$product->description = $validated['description'] ?? null;
 		$product->prix = $validated['prix'] ?? null;
 		$product->type_paiement = $validated['type_paiement'] ?? null;
@@ -128,14 +201,50 @@ class ProductController extends Controller
 		$product->category_id = $validated['category_id'] ?? null;
 		$product->municipality_id = $validated['municipality_id'] ?? null;
 		$product->user_id = $validated['user_id'] ?? null;
-		
+
         $product->save();
 
         $data = [
             'success'       => true,
             'product'   => $product
         ];
-        
+
+        return response()->json($data);
+    }
+
+    public function user_update(UpdateProductRequest $request, Product $product)
+    {
+        $user = Auth::getUser($request, Auth::USER);
+        $validated = $request->validated();
+
+        if ($user->id !== $product->user_id) {
+            $data = [
+                'error' => true,
+                'message' => 'forbidden'
+            ];
+
+            return response()->json($data, 403);
+        }
+
+        $product->nom = $validated['nom'] ?? null;
+		$product->slug = Str::slug($validated['slug']);
+		$product->description = $validated['description'] ?? null;
+		$product->prix = $validated['prix'] ?? null;
+		$product->type_paiement = $validated['type_paiement'] ?? null;
+		$product->type = $validated['type'] ?? null;
+		$product->display_img_url_list = $validated['display_img_url_list'] ?? null;
+		$product->images_url_list = $validated['images_url_list'] ?? null;
+		$product->category_id = $validated['category_id'] ?? null;
+		$product->municipality_id = $validated['municipality_id'] ?? null;
+		$product->user_id = $user->id;
+
+        $product->save();
+
+        $data = [
+            'success'       => true,
+            'product'   => $product
+        ];
+
         return response()->json($data);
     }
 
@@ -146,7 +255,30 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Product $product)
-    {   
+    {
+        $product->delete();
+
+        $data = [
+            'success' => true,
+            'product' => $product
+        ];
+
+        return response()->json($data);
+    }
+
+    public function user_destroy(Request $request, Product $product)
+    {
+        $user = Auth::getUser($request, Auth::USER);
+
+        if ($product->user_id !== $user->id ) {
+            $data = [
+                'error' => true,
+                'message' => 'forbidden'
+            ];
+
+            return response()->json($data, 403);
+        }
+
         $product->delete();
 
         $data = [
